requirements.txt
requests
python-dotenv
google-genai
elevenlabs
fpdf2
youdotcom
feedparser
streamlit
edge-tts
yfinance
plotly
nest-asyncio
trafilatura
beautifulsoup4
